Remove logging refinement
I can't get this to work at all.  I suspect that things are instantiated
before the refinement is loaded, or something.  I've spent a couple of
hours trying to get this going to no avail.  So whilst I'd really like
to have this work, emitting a log line each time the auth bypass is
used, I can't make it go :/